登陆处理

+++ src/main/java/com/inca/mapper/UserMapper.java
@@ -8,6 +8,6 @@ import com.inca.entity.pub.ex.UserExample;
 @Repository
 public interface UserMapper {
 	User selectUserById(Integer id);
-	Integer countUser(UserExample ue);
+	Long countUser(UserExample ue);
 
 }
